refactor: App.tsx에서 Jotai Provider와 AppContent 컴포넌트로 구조 변경 및 알림과 관리자 대시보드 관련 로직을 AppContent로 이동
